+++ wand/api.py
@@ -292,6 +292,11 @@ try:
 
     library.MagickSetImageColorspace.argtypes = [ctypes.c_void_p, ctypes.c_int]
 
+    library.MagickGetImageCompression.argtypes = [ctypes.c_void_p]
+    library.MagickGetImageCompression.restype = ctypes.c_int
+
+    library.MagickSetImageCompression.argtypes = [ctypes.c_void_p, ctypes.c_int]
+
     library.MagickGetImageDepth.argtypes = [ctypes.c_void_p]
     library.MagickGetImageDepth.restype = ctypes.c_size_t