+++ docs/changes.rst
@@ -1,6 +1,14 @@
 Wand Changelog
 ==============
 
+Version 0.3.7
+-------------
+
+Released on Match 25, 2014.
+
+- A hotfix of debug prints made at 0.3.6.
+
+
 Version 0.3.6
 -------------